Combine ModbusClientProtocol and ModbusUdpClientProtocol.
